Starting point for cleaning up javascript
